More warnings on 32-bit Linux/gcc

+++ src/Array.h
@@ -121,21 +121,21 @@ public:
 
 	template <class F> size_t Find(F function_, int64_t value, size_t start, size_t end) const {
 		const F function = {};
-		if(end == -1)
+		if(end == (size_t)-1)
 			end = m_len;
 		for(size_t s = start; s < end; s++) {
 			if(function(value, Get(s)))
 				return s;
 		}
-		return -1;
+		return (size_t)-1;
 	}
 	void Preset(int64_t min, int64_t max, size_t count);
 	void Preset(size_t bitwidth, size_t count); 
 	void FindAll(Array& result, int64_t value, size_t offset=0, size_t start=0, size_t end=(size_t)-1) const;
 	void FindAllHamming(Array& result, uint64_t value, size_t maxdist, size_t offset=0) const;
-	int64_t Sum(size_t start = 0, size_t end = -1) const;
-	bool Max(int64_t& result, size_t start = 0, size_t end = -1) const;
-	bool Min(int64_t& result, size_t start = 0, size_t end = -1) const;
+	int64_t Sum(size_t start = 0, size_t end = (size_t)-1) const;
+	bool Max(int64_t& result, size_t start = 0, size_t end = (size_t)-1) const;
+	bool Min(int64_t& result, size_t start = 0, size_t end = (size_t)-1) const;
 	template <class F> size_t Query(int64_t value, size_t start, size_t end);
 
 	void Sort(void);
